+++ app/src/main/java/app/revanced/integrations/twitter/settings/SettingsAboutFragment.java
@@ -12,6 +12,7 @@ import app.revanced.integrations.shared.Utils;
 import com.twitter.ui.widget.LegacyTwitterPreferenceCategory;
 import app.revanced.integrations.BuildConfig;
 import java.util.*;
+import app.revanced.integrations.shared.StringRef;
 
 @SuppressWarnings("deprecation")
 public class SettingsAboutFragment extends PreferenceFragment implements Preference.OnPreferenceClickListener {
@@ -60,21 +61,21 @@ public class SettingsAboutFragment extends PreferenceFragment implements Prefere
         flags.put(strRes("custom_navigation"),SettingsStatus.navBarCustomisation);
         flags.put(strRes("piko_pref_download"),SettingsStatus.changeDownloadEnabled);
         flags.put(strRes("piko_pref_download_media_link_handle"),SettingsStatus.mediaLinkHandle);
-        flags.put(strRes("piko_pref_hide_promoted_posts"),SettingsStatus.hideAds);
-        flags.put(strRes("piko_pref_hide_g_ads"),SettingsStatus.hideGAds);
-        flags.put(strRes("piko_pref_wtf_section"),SettingsStatus.hideWTF);
-        flags.put(strRes("piko_pref_cts_section"),SettingsStatus.hideCTS);
-        flags.put(strRes("piko_pref_ctj_section"),SettingsStatus.hideCTJ);
-        flags.put(strRes("piko_pref_ryb_section"),SettingsStatus.hideRBMK);
-        flags.put(strRes("piko_pref_pinned_posts_section"),SettingsStatus.hideRPinnedPosts);
-        flags.put(strRes("piko_pref_hide_detailed_posts"),SettingsStatus.hideDetailedPosts);
-        flags.put(strRes("piko_pref_hide_trends"),SettingsStatus.hidePromotedTrend);
+        flags.put(strRemoveRes("piko_pref_hide_promoted_posts"),SettingsStatus.hideAds);
+        flags.put(strRemoveRes("piko_pref_hide_g_ads"),SettingsStatus.hideGAds);
+        flags.put(strRemoveRes("piko_pref_wtf_section"),SettingsStatus.hideWTF);
+        flags.put(strRemoveRes("piko_pref_cts_section"),SettingsStatus.hideCTS);
+        flags.put(strRemoveRes("piko_pref_ctj_section"),SettingsStatus.hideCTJ);
+        flags.put(strRemoveRes("piko_pref_ryb_section"),SettingsStatus.hideRBMK);
+        flags.put(strRemoveRes("piko_pref_pinned_posts_section"),SettingsStatus.hideRPinnedPosts);
+        flags.put(strRemoveRes("piko_pref_hide_detailed_posts"),SettingsStatus.hideDetailedPosts);
+        flags.put(strRemoveRes("piko_pref_hide_trends"),SettingsStatus.hidePromotedTrend);
         flags.put(strRes("piko_pref_chirp_font"),SettingsStatus.enableFontMod);
-        flags.put(strRes("piko_pref_hide_fab"),SettingsStatus.hideFAB);
-        flags.put(strRes("piko_pref_hide_fab_menu"),SettingsStatus.hideFABBtns);
+        flags.put(strRemoveRes("piko_pref_hide_fab"),SettingsStatus.hideFAB);
+        flags.put(strRemoveRes("piko_pref_hide_fab_menu"),SettingsStatus.hideFABBtns);
         flags.put(strRes("piko_pref_show_sensitive_media"),SettingsStatus.showSensitiveMedia);
         flags.put(strRes("piko_pref_selectable_text"),SettingsStatus.selectableText);
-        flags.put(strRes("piko_pref_rec_users"),SettingsStatus.hideRecommendedUsers);
+        flags.put(strRemoveRes("piko_pref_rec_users"),SettingsStatus.hideRecommendedUsers);
         flags.put(strRes("piko_pref_browser_chooser"),SettingsStatus.browserChooserEnabled);
         flags.put(strRes("piko_pref_custom_share_domain"),SettingsStatus.customSharingDomainEnabled);
         flags.put(strRes("piko_pref_feature_flags"),SettingsStatus.featureFlagsEnabled);
@@ -83,21 +84,23 @@ public class SettingsAboutFragment extends PreferenceFragment implements Prefere
         flags.put(strRes("piko_pref_customisation_navbartabs"),SettingsStatus.navBarCustomisation);
         flags.put(strRes("piko_pref_customisation_sidebartabs"),SettingsStatus.sideBarCustomisation);
         flags.put(strRes("piko_pref_disable_auto_timeline_scroll"),SettingsStatus.disableAutoTimelineScroll);
-        flags.put(strRes("piko_pref_hide_live_threads"),SettingsStatus.hideLiveThreads);
-        flags.put(strRes("piko_pref_hide_banner"),SettingsStatus.hideBanner);
-        flags.put(strRes("piko_pref_hide_bmk_timeline"),SettingsStatus.hideInlineBmk);
+        flags.put(strRemoveRes("piko_pref_hide_live_threads"),SettingsStatus.hideLiveThreads);
+        flags.put(strRemoveRes("piko_pref_hide_view_count"),SettingsStatus.hideViewCount);
+        flags.put(strRemoveRes("piko_pref_hide_banner"),SettingsStatus.hideBanner);
+        flags.put(strRemoveRes("piko_pref_hide_bmk_timeline"),SettingsStatus.hideInlineBmk);
         flags.put(strRes("piko_pref_show_poll_result"),SettingsStatus.showPollResultsEnabled);
-        flags.put(strRes("piko_pref_comm_notes"),SettingsStatus.hideCommunityNote);
-        flags.put(strRes("piko_pref_hide_quick_promote"),SettingsStatus.hidePromoteButton);
-        flags.put(strRes("piko_pref_hide_immersive_player"),SettingsStatus.hideImmersivePlayer);
+        flags.put(strRemoveRes("community_notes_title"),SettingsStatus.hideCommunityNote);
+        flags.put(strRemoveRes("piko_pref_hide_quick_promote"),SettingsStatus.hidePromoteButton);
+        flags.put(strRemoveRes("piko_pref_hide_immersive_player"),SettingsStatus.hideImmersivePlayer);
         flags.put(strRes("piko_pref_clear_tracking_params"),SettingsStatus.cleartrackingparams);
         flags.put(strRes("piko_pref_unshorten_link"),SettingsStatus.unshortenlink);
         flags.put(strRes("piko_pref_force_translate"),SettingsStatus.forceTranslate);
         flags.put(strRes("piko_pref_round_off_numbers"),SettingsStatus.roundOffNumbers);
         flags.put(strRes("piko_pref_customisation_inlinetabs"),SettingsStatus.inlineBarCustomisation);
         flags.put(strRes("piko_pref_debug_menu"),SettingsStatus.enableDebugMenu);
-        flags.put(strRes("piko_pref_hide_premium_prompt"),SettingsStatus.hidePremiumPrompt);
-        flags.put(strRes("piko_pref_hide_hidden_replies"),SettingsStatus.hideHiddenReplies);
+        flags.put(strRemoveRes("piko_pref_hide_premium_prompt"),SettingsStatus.hidePremiumPrompt);
+        flags.put(strRemoveRes("piko_pref_hide_hidden_replies"),SettingsStatus.hideHiddenReplies);
+        flags.put(strRes("piko_pref_del_from_db"),SettingsStatus.deleteFromDb);
 
         LegacyTwitterPreferenceCategory patPref = preferenceCategory(strRes("piko_pref_patches"), screen);
 
@@ -160,7 +163,9 @@ public class SettingsAboutFragment extends PreferenceFragment implements Prefere
     }
 
     private static String strRes(String tag) {
-        return app.revanced.integrations.twitter.Utils.strRes(tag);
+        return StringRef.str(tag);
     }
-
+    private static String strRemoveRes(String tag) {
+        return StringRef.str("piko_pref_remove",strRes(tag));
+    }
 }